Added setup for mvx-esdt-safe interact in separate folder
